M12258: Further explanation requiered
Hello, @apwestgarth,

The source states to substitute your domain for "azurestack.local". However, the form mentioned is "https://portal.local.azurestack.external/tokenauthorize", i.e. this form does not contain "azurestack.local".
In addition, it is not quite clear what is to be replaced: the domain or "azurestack.local"? It seems that "azurestack.local" is replaced by the domain.

Please, help to check my comment tag added into the article and reply with explanation if source text should be fixed, or it is by design. If fix is needed, please, let me know either if you would like me to fix it in this PR, or if you will fix it in another PR. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR. 

Many thanks in advance.