M29530: Possible incomplete sentence at line 182
Hello, @brenduns,  
Translator has reported possible source content issue. 
"The sentence seems to be incomplete."
Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.
Many thanks in advance.