M31,073: Missing verb
@mattbriggs
Translator has reported possible source content issue. 
Desciption: A verb is missing. We assumed as follows: "You can create measures in the portal, get them from the REST API, or query them with PowerShell or CLI.".
Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR.