Updating release notes
Commented out a fix as it actually didn't go in 1910 due to build issues, it will be in the next build (2002), so commenting out until the next build.

Commented out the following fix:
Fixed an issue where adding more than one public IP on the same NIC on a Virtual Machine resulted in internet connectivity issues. Now, a NIC with two public IPs should work as expected.