Removing a line
Removing - "If you're using the ASDK, this VM is named AzS-ERCS01. Update the `<placeholder>` values before running:" because it insinuates that secret rotation can be performed on an ASDK, which is not true. 
ASDK does not support secret rotation. 

Updating based on customer's question on forum - https://docs.microsoft.com/en-us/answers/questions/310452/index.html