Update azure-stack-disconnected-deployment.md
Adding clarity on the "Certificates" item in the table of impaired items for Disconnected stamps. Amending from "Certificates" to "Certificates issued by a public certificate authority (CA)”. And I have also added an additional sentence in the description to make it cleat that fully disconnected, customers should use a private Certificate Authority (CA) server that is within their organization.