Update deploy-arc-resource-bridge-using-command-line.md
Removing backslash at end of example of csv_path.  It results in an issue when creating VM from Azure Arc it fails to attach VHDX to VM, this seems to cause error 0x80070035 "An unexpected error occurred: The network path was not found." because there will be a \\ in the imagestore path.