Update aks-networks.md
I have added several examples in the variables for the creation of the AKS network to make it easier to understand. In particular the variable $ipaddressprefix requires the network with the subfix CIDR at the end of it and if you don't provide an example it is difficult to visualize it.