Update deployment-install-os.md
Hi all, I don't know if this is the best place to indicate this information, but the powershell modules required to perform the installation (the registration of the nodes in ARC) are not present in the Azure Stack HCI image and these have to be installed. 
In case I am wrong, I would love to know where this step is documented. 

Thank you very much in advance, 
Cristian