Update whats-new-2306-preview.md
added remark that Supplemental package is not supported and also not allowed as a base to upgrade from 22H2. With this the user should be made aware to reconsider attempts using the supplemental package.

As of October 2024, there is unfortunately, this a use case for Azure Stack HCI 22H2 new or re-installations. Such as Stretched Clusters or customers that cannot accept the current limitations with their BCDR solutions.