Update aks-windows-server-retirement.md
Removed "There are no separate costs for running AKS clusters, including Linux and Windows node pools." from the Pricing statement in the "What is AKS on Azure Local?" section. Because the statement is not accurate which has caused some confusion to customers and field teams. Actually, customers need to take care of the Windows license separately for using Windows node pools.