Update oem-license.md - instance vs. clusters
changed most of "clusters" into instances, now that is the correct term. In Azure Portal you can only create Azure Local instances, not clusters. Single Node does not form a "cluster" by design but also an instance. So it is a more universal term chosen by PG. Also it honours it is a unified solution.

This PR shall reflect these changes in the product language used.